# api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Shopmon API
  description: Shopware monitoring application API
  version: 1.0.0
servers:
  - url: /api
    description: API base path
...
security:
  - bearerAuth: []
paths:
...
  /account/me:
    get:
      operationId: getAccountMe
      tags: [Account]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/extensions:
    get:
      operationId: getAccountExtensions
      tags: [Account]
      summary: Get aggregated extensions across all shops
      responses:
        "200":
          description: List of extensions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountExtension"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/organizations:
    get:
      operationId: getAccountOrganizations
      tags: [Account]
      summary: Get organizations the user belongs to
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountOrganization"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/shops:
    get:
      operationId: getAccountShops
      tags: [Account]
      summary: Get all shops accessible to the user
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountShop"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/projects:
    get:
      operationId: getAccountProjects
      tags: [Account]
      summary: Get all projects accessible to the user
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountProject"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/changelogs:
    get:
      operationId: getAccountChangelogs
      tags: [Account]
      summary: Get changelogs across all shops
      responses:
        "200":
          description: List of changelogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountChangelog"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/subscribed-shops:
    get:
      operationId: getAccountSubscribedShops
      tags: [Account]
      summary: Get shops the user is subscribed to for notifications
      responses:
        "200":
          description: List of subscribed shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscribedShop"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  # Notifications
  # ──────────────────────────────────────────────
...
  /auth/exchange-code:
    post:
      operationId: exchangeCode
      tags: [Auth]
      summary: Exchange one-time authorization code for session token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: Invalid or expired code
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication

  # ──────────────────────────────────────────────
  # Parameters
  # ──────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    # ── Account ──────────────────────────────────
    UserProfile:
      type: object
      required:
        - id
        - displayName
        - email
        - createdAt
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    AccountExtension:
      type: object
      required:
        - name
        - label
        - version
        - latestVersion
        - active
        - installed
        - storeLink
        - ratingAverage
        - installedAt
        - changelog
        - shops
      properties:
        name:
          type: string
        label:
          type: string
        version:
          type: string
        latestVersion:
          type: string
        active:
          type: boolean
        installed:
          type: boolean
        storeLink:
          type: string
          nullable: true
        ratingAverage:
          type: number
          nullable: true
        installedAt:
          type: string
          nullable: true
          format: date-time
        changelog:
          type: string
          nullable: true
        shops:
          type: array
          items:
            $ref: "#/components/schemas/AccountExtensionShop"
    AccountExtensionShop:
      type: object
      required:
        - shopId
        - shopName
        - shopOrganizationName
        - shopOrganizationId
        - shopUrl
        - active
        - version
        - latestVersion
        - installed
      properties:
        shopId:
          type: integer
        shopName:
          type: string
        shopOrganizationName:
          type: string
        shopOrganizationId:
          type: string
        shopUrl:
          type: string
        active:
          type: boolean
        version:
          type: string
        latestVersion:
          type: string
        installed:
          type: boolean
    AccountOrganization:
      type: object
      required:
        - id
        - name
        - logo
        - createdAt
        - shopCount
        - memberCount
      properties:
        id:
          type: string
        name:
          type: string
        logo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        shopCount:
          type: integer
        memberCount:
          type: integer
    AccountShop:
      type: object
      required:
        - id
        - name
        - url
        - favicon
        - status
        - shopwareVersion
        - lastScrapedAt
        - lastScrapedError
        - organizationId
        - organizationName
        - projectId
        - projectName
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        favicon:
          type: string
          nullable: true
        status:
          type: string
        shopwareVersion:
          type: string
        lastScrapedAt:
          type: string
          nullable: true
          format: date-time
        lastScrapedError:
          type: string
          nullable: true
        organizationId:
          type: string
        organizationName:
          type: string
        projectId:
          type: integer
          nullable: true
        projectName:
          type: string
          nullable: true
    AccountProject:
      type: object
      required:
        - id
        - name
        - description
        - gitUrl
        - organizationId
        - organizationName
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        gitUrl:
          type: string
          nullable: true
        organizationId:
          type: string
        organizationName:
          type: string
    AccountChangelog:
      type: object
      required:
        - id
        - shopId
        - extensions
        - oldShopwareVersion
        - newShopwareVersion
        - date
        - shopName
        - shopOrganizationName
        - shopOrganizationId
      properties:
        id:
          type: integer
        shopId:
          type: integer
        extensions:
          type: array
          description: JSON data describing extension changes
          items:
            $ref: "#/components/schemas/ExtensionDiff"
        oldShopwareVersion:
          type: string
          nullable: true
        newShopwareVersion:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        shopName:
          type: string
        shopOrganizationName:
          type: string
        shopOrganizationId:
          type: string
    ExtensionDiff:
      type: object
      required:
        - name
        - label
        - state
        - active
      properties:
        name:
          type: string
        label:
          type: string
        state:
          type: string
        oldVersion:
          type: string
          nullable: true
        newVersion:
          type: string
          nullable: true
        changelog:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ExtensionChangelogEntry"
        active:
          type: boolean
    ExtensionChangelogEntry:
      type: object
      required:
        - version
        - text
        - creationDate
        - isCompatible
      properties:
        version:
          type: string
        text:
          type: string
        creationDate:
          type: string
          format: date-time
        isCompatible:
          type: boolean
    SubscribedShop:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string

    # ── Notifications ────────────────────────────
...